pyproject.toml
[project]
name = "reports-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.49.0",
    "watchdog>=6.0.0",
]
